Instance normalization: enabled affine transformation by default
BatchNorm has a default affine=True, which makes bias term unnecessary,
but the InstanceNorm default is affine=False, because of historical reasons.
While it is an open issue in Pytorch, see pytorch/pytorch#22755,
it would be better to make it explicit. (Let's assume the defaults change, it is better
to be explicit. Otherwise line 314 could be also simply changed to 
"use_bias = not batch_norm" )

Long story short: InstanceNorm needs bias term, and i think the PR is a reasonable solution.
(But obviously test it.) In case of accepted PR, a new Pypi release would be nice too.